Fix usage of uninitialized value in libupnp
Although the documentation says that the SubsId parameter of UpnpSubscribe() is
out-only, the library in fact accesses it for reading before any writing (clones
the current value with strdup). According to the documentation it would be safe
to pass a pointer to an uninitialized memory area to UpnpSubscribe, but reality
disproves it.