Fix Crypto++ detection
The original detection code was flawed in several aspects. When it came to
cross-compiling, it used to find the build system's cryptopp installation
unless a prefix was given. Then it screwed up everything, by including build
flags for host.

The current approach doesn't use any hard-coded directories. It lets the compiler
find cryptopp, and then we ask the compiler where it found it, to let us check
the version number (as cryptopp still doesn't have any compile-time or run-time
version information). If a prefix was given, then we test the given prefix for
something that looks like cryptopp, and live with that if that seems to be ok.

We don't anymore set CPPFLAGS nor LDFLAGS for cryptopp installations found in
system directories (i.e. reachable for the compiler without extra flags), we set
them only if --with-crypto-prefix was given and it points to an installed version
of the library.